Generalizing the styler
